SSCS-12148 - Clear specific body content fields before about to submit post hearing request response returned
