+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/posthearingrequest/PostHearingRequestAboutToSubmitHandler.java
@@ -12,10 +12,7 @@ import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
 import uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType;
 import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
-import uk.gov.hmcts.reform.sscs.ccd.domain.DocumentLink;
-import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
-import uk.gov.hmcts.reform.sscs.ccd.domain.PostHearingRequestType;
-import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
+import uk.gov.hmcts.reform.sscs.ccd.domain.*;
 import uk.gov.hmcts.reform.sscs.ccd.presubmit.PreSubmitCallbackHandler;
 import uk.gov.hmcts.reform.sscs.service.FooterService;
 import uk.gov.hmcts.reform.sscs.util.PdfRequestUtil;
@@ -58,6 +55,8 @@ public class PostHearingRequestAboutToSubmitHandler implements PreSubmitCallback
                 callback.getEvent());
         }
 
+        clearPostHearingRequestFormatAndContentFields(caseData, caseData.getPostHearing().getRequestType());
+
         return response;
     }
 
@@ -89,6 +88,36 @@ public class PostHearingRequestAboutToSubmitHandler implements PreSubmitCallback
                 .documentHash(previewDocument.getDocumentHash())
                 .build();
             caseData.getDocumentStaging().setPreviewDocument(renamedPreviewDoc);
+        }
+    }
+
+    private static void clearPostHearingRequestFormatAndContentFields(SscsCaseData caseData, PostHearingRequestType requestType) {
+        PostHearing postHearing = caseData.getPostHearing();
+        DocumentGeneration docGen = caseData.getDocumentGeneration();
+
+        switch (requestType) {
+            case SET_ASIDE -> {
+                postHearing.getSetAside().setRequestFormat(null);
+                docGen.setBodyContent(null);
+            }
+            case CORRECTION -> {
+                postHearing.getCorrection().setRequestFormat(null);
+                docGen.setCorrectionBodyContent(null);
+            }
+            case STATEMENT_OF_REASONS -> {
+                postHearing.getStatementOfReasons().setRequestFormat(null);
+                docGen.setStatementOfReasonsBodyContent(null);
+            }
+            case LIBERTY_TO_APPLY -> {
+                postHearing.getLibertyToApply().setRequestFormat(null);
+                docGen.setLibertyToApplyBodyContent(null);
+            }
+            case PERMISSION_TO_APPEAL -> {
+                postHearing.getPermissionToAppeal().setRequestFormat(null);
+                docGen.setPermissionToAppealBodyContent(null);
+            }
+            default -> {
+            }
         }
     }
 }
